:tada: merge OpenVAS XML and CSV parsers (#9322)
* :tada: merge OpenVAS XML and CSV formats to have only one parser for OpenVAS

* Update docs/content/en/integrations/parsers/file/openvas.md


---------
