+++ docs/content/en/integrations/parsers/file/openvas.md
@@ -1,0 +1,5 @@
+---
+title: "OpenVAS Parser"
+toc_hide: true
+---
+You can either upload the exported results of an OpenVAS Scan in a .csv or .xml format.
