--- docs/content/en/integrations/parsers/file/openvas_csv.md
@@ -1,5 +1,0 @@
----
-title: "OpenVAS CSV"
-toc_hide: true
----
-Import OpenVAS Scan in CSV format. Export as CSV Results on OpenVAS.
