--- docs/content/en/integrations/parsers/file/openvas_xml.md
@@ -1,5 +1,0 @@
----
-title: "OpenVAS XML"
-toc_hide: true
----
-Import Greenbone OpenVAS Scan in XML format. Export as XML Results on OpenVAS.
