+++ unittests/tools/test_openvas_parser.py
@@ -1,16 +1,15 @@
 from ..dojo_test_case import DojoTestCase
-from dojo.tools.openvas_csv.parser import OpenVASCsvParser
+from dojo.tools.openvas.parser import OpenVASParser
 from dojo.models import Test, Engagement, Product
 
 
-class TestOpenVASUploadCsvParser(DojoTestCase):
-
+class TestOpenVASParser(DojoTestCase):
     def test_openvas_csv_one_vuln(self):
-        with open("unittests/scans/openvas_csv/one_vuln.csv") as f:
+        with open("unittests/scans/openvas/one_vuln.csv") as f:
             test = Test()
             test.engagement = Engagement()
             test.engagement.product = Product()
-            parser = OpenVASCsvParser()
+            parser = OpenVASParser()
             findings = parser.get_findings(f, test)
             for finding in findings:
                 for endpoint in finding.unsaved_endpoints:
@@ -27,11 +26,11 @@ class TestOpenVASUploadCsvParser(DojoTestCase):
             self.assertEqual(22, findings[0].unsaved_endpoints[0].port)
 
     def test_openvas_csv_many_vuln(self):
-        with open("unittests/scans/openvas_csv/many_vuln.csv") as f:
+        with open("unittests/scans/openvas/many_vuln.csv") as f:
             test = Test()
             test.engagement = Engagement()
             test.engagement.product = Product()
-            parser = OpenVASCsvParser()
+            parser = OpenVASParser()
             findings = parser.get_findings(f, test)
             for finding in findings:
                 for endpoint in finding.unsaved_endpoints:
@@ -48,3 +47,40 @@ class TestOpenVASUploadCsvParser(DojoTestCase):
             self.assertEqual("LOGSRV", endpoint.host)
             self.assertEqual("tcp", endpoint.protocol)
             self.assertEqual(9200, endpoint.port)
+
+    def test_openvas_xml_no_vuln(self):
+        with open("unittests/scans/openvas/no_vuln.xml") as f:
+            test = Test()
+            test.engagement = Engagement()
+            test.engagement.product = Product()
+            parser = OpenVASParser()
+            findings = parser.get_findings(f, test)
+            self.assertEqual(0, len(findings))
+
+    def test_openvas_xml_one_vuln(self):
+        with open("unittests/scans/openvas/one_vuln.xml") as f:
+            test = Test()
+            test.engagement = Engagement()
+            test.engagement.product = Product()
+            parser = OpenVASParser()
+            findings = parser.get_findings(f, test)
+            for finding in findings:
+                for endpoint in finding.unsaved_endpoints:
+                    endpoint.clean()
+            self.assertEqual(1, len(findings))
+            with self.subTest(i=0):
+                finding = findings[0]
+                self.assertEqual("Mozilla Firefox Security Update (mfsa_2023-32_2023-36) - Windows_10.0.101.2_general/tcp", finding.title)
+                self.assertEqual("Critical", finding.severity)
+
+    def test_openvas_xml_many_vuln(self):
+        with open("unittests/scans/openvas/many_vuln.xml") as f:
+            test = Test()
+            test.engagement = Engagement()
+            test.engagement.product = Product()
+            parser = OpenVASParser()
+            findings = parser.get_findings(f, test)
+            for finding in findings:
+                for endpoint in finding.unsaved_endpoints:
+                    endpoint.clean()
+            self.assertEqual(44, len(findings))
